+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.3.56
+- Added `operationId` in `AtsEntry`
+
+## 3.3.55
+
+- Added lock, suspension and paymentWarning fields to `User` model
+
 ## 3.3.54
 - updated `TriggerInput` model to mathch the new `Trigger` model